Bug 1808489: Fix for sort/filter on helm details resources page
